requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
pyodbc>=4.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
